Phase 1: retire packages/game-generator and its dead endpoint
The last Phase 1 item. GameProject moves to @gamedevpl/contract, and the
generator seam it was living in goes away entirely: -2028/+64 across 52 files.

POST /api/generate-game was NOT dev-only
It shipped in production, behind a session, a 30/hour per-IP rate limit, a paid
Vertex moderation call and a daily quota — but with zero callers. The web threaded
`onGenerateMock` from App.tsx through CreatePage into HeroPromptSection, which
never destructured it and never called it. `grep -rn "onGenerateMock("` returns
nothing. So `handleGenerateMock` was unreachable, `mockStatus` could never leave
'idle', `mockError` could never be set, and the `{ type: 'generated' }` stage
branch was dead. Removing it strictly reduces attack and cost surface.

GET /api/health keeps its exact shape
`provider` stayed in the response rather than being dropped with the generator it
described, so any external monitor keying off it keeps working. It is now a fixed
'mock' with a comment saying why. app.test.ts still asserts the full shape.

Removed with it, because nothing else read them
- apps/api/src/generator.ts (createGenerator; only ever accepted 'mock')
- apps/web/src/api.ts (generateGame; the file had no other export)
- dailyGenerationQuota + DAILY_GENERATION_QUOTA — only ever gated this route
- LLM_PROVIDER, and the apps/api/.env.example container block that was explicitly
  headed "Only read when LLM_PROVIDER=container". Zero code references to any of
  LLM_PROVIDER / AGENT_MODE / AGENT_CMD / AGENT_ARGS / AGENT_TIMEOUT_MS /
  AGENT_RUNNER_IMAGE. That block also cited packages/game-generator/src/container.ts,
  a file that has not existed for some time.
- the src/engine/*.js trio (683 lines), whose only consumer was a test asserting
  the files exist

store.ts keeps its `mocks` usage counter: it is persisted Firestore data, and the
file is frozen.

Two guard tests earned their keep
- moderation-metrics.test.ts enumerates every module that calls
  logModerationRejection and failed the moment app.ts stopped being one. Expected
  list updated.
- The comment-prose ratchet caught a subtler thing: deleting the mockStatus/
  mockError props left their comment stranded directly above an unrelated one,
  and two adjacent // lines count as a prose stack. The stale comment is gone.

Lockfile
Removed the three game-generator stanzas by hand. Regenerating from scratch was
tried first and rejected: it dropped 1710 lines and 49 packages, pruning
platform-specific optional dependencies down to this machine's platform, which
would have broken `npm ci` on CI's. The surgical edit is -13 lines and npm
reports the tree up to date.

Docs
Updated the agent-facing and index docs that would now point at a package that
does not exist: AGENTS.md, .github/copilot-instructions.md,
docs/contributing-for-agents.md, docs/README.md, docs/local-development.md.
Left alone: architecture.md, roadmap.md, auth-and-usage-plan.md,
content-safety-plan.md, games-repo-blueprint.md, how-to-play-plan.md and
container-orchestration.md still mention the route. They are superseded plan
docs already named in debt D8, and belong to the docs-truth workstream rather
than to this change.

Templates
docs/games-repo-blueprint.md maps the three templates to dodge-the-falling-rocks,
collect-the-coins and fly-through-asteroids, and docs/steel-thread-plan.md records
that scaffold as merged into the games repo — so they are already seeded there and
are not being lost here.
