+++ .env.example
@@ -36,7 +36,6 @@
 # BETA_ALLOWED_UIDS=
 # BETA_ALLOWED_EMAILS=
 # ADMIN_UIDS=
-# DAILY_GENERATION_QUOTA=20
 # DAILY_REFINE_QUOTA=20
 # Automation ceiling for bot: accounts — the deploy gate spends one refine per deploy.
 # DAILY_REFINE_QUOTA_BOT=200
@@ -61,7 +60,6 @@
 # RESEND_API_KEY=              # transactional email
 # VAPID_PUBLIC_KEY=            # web push
 # PROJECT_ID=                  # Firestore (production only; dev uses an in-memory store)
-# LLM_PROVIDER=mock            # only "mock" is supported
 # ENABLE_VERTEX_MODERATION=    # content moderation via Vertex
 # VERTEX_PROJECT_ID=
 # VERTEX_REGION=
