+++ .github/copilot-instructions.md
@@ -14,7 +14,7 @@ file is the short actionable summary.
 ## What this is
 
 - Monorepo, **npm workspaces**: `apps/web` (Vite + React + TS), `apps/api` (Fastify + TS),
-  `packages/game-generator` (local preview seam), `infra/` (placeholder), `docs/` (docs).
+  `packages/contract` (shared types), `packages/zone-core`, `infra/` (placeholder), `docs/` (docs).
 - The deterministic mock proves the local player loop. Production games are maintained by
   coding agents in a dedicated games repository and published as static bundles; do not replace
   the mock with an in-app agent runtime.
