+++ apps/api/.env.example
@@ -1,43 +1,3 @@
-# Which game generator to use.
-#   mock      — deterministic, offline, zero cost (default)
-#   container — runs the agent-runner container image (see containers/agent-runner)
-LLM_PROVIDER=mock
-
 # Server binding
 PORT=3001
 HOST=127.0.0.1
-
-# ─── container provider ──────────────────────────────────────────────────────
-# Only read when LLM_PROVIDER=container.
-
-# Image to run. Build it with: npm run agent-runner:build
-# AGENT_RUNNER_IMAGE=gamedevpl/agent-runner:latest
-
-# What runs INSIDE the container:
-#   mock     — fills the bundled template offline. No network, no tokens. (default)
-#   external — runs a real coding-agent CLI. Costs real money per run.
-# AGENT_MODE=mock
-
-# ─── external mode: real agent ───────────────────────────────────────────────
-# Unset by default on purpose — with no AGENT_CMD, external mode refuses to run,
-# so nothing can spend money by accident.
-#
-# Claude Code CLI (installed in the image) — the prompt arrives on stdin, and
-# acceptEdits is what lets it write files without an interactive approval prompt:
-# AGENT_CMD=claude
-# AGENT_ARGS=["-p","--permission-mode","acceptEdits"]
-# AGENT_TIMEOUT_MS=600000
-
-# Credentials for the agent. These are forwarded into the container BY NAME, so
-# the value never appears in the docker command line. Put the real value in
-# .env (gitignored) — never commit it.
-# ANTHROPIC_API_KEY=
-# OPENAI_API_KEY=
-
-# ─── auth proxy (required for AGENT_MODE=external) ───────────────────────────
-# The container is never given a real provider key. It gets a short-lived,
-# job-scoped token instead, and talks to the proxy rather than the provider.
-# Run the proxy with: npm run dev -w @gamedevpl/auth-proxy
-#
-# AUTH_PROXY_URL=http://host.docker.internal:3002
-# AUTH_PROXY_SECRET=<random string; must match the proxy's own AUTH_PROXY_SECRET>
