+++ docs/README.md
@@ -37,7 +37,7 @@ survives only in this repo's early history.
 | Multiplayer     | ✅ Party mode — one shared screen, phones as controllers (relay is in-process, hence one instance — splitting it out is built, see `deployment.md`)                                                                                                                 |
 | Mobile          | ✅ Every catalog game is playable with a thumb, **enforced in CI** from each game's source. No PWA/install path yet                                                                                                                                                 |
 | Device sensing  | 🚧 Tilt play shipped; camera backdrop in flight — the shell reads sensors / `getUserMedia` on its own origin (`apps/web/src/sensing.ts`; games-repo `docs/camera-ar-platform.md`). Frames and raw readings never leave the browser; the iframe sandbox is unchanged |
-| Generator seam  | `packages/game-generator` — deterministic **mock only**, a development preview route                                                                                                                                                                                |
+| Shared contract | `packages/contract` — route tables, status vocabularies, response shapes, protocol versions                                                                                                                                                                         |
 | Local dev       | ✅ Whole product runs with no keys — bundled fixture games and a dev sign-in (`local-development.md`)                                                                                                                                                               |
 | Legal           | ✅ Terms, privacy policy, AI disclosure, and a DSA notice-and-action route are published                                                                                                                                                                            |
 | Orchestration   | 🗃️ Self-hosted agent execution abandoned for legal reasons (see `games-repo.md`); ✅ hosted Anthropic managed builder (MCP + per-round vault) is live when valid configuration selects it through the environment registry (`managed-agent-backend.md`)             |
