+++ eslint-rules/comment-prose-baseline.json
@@ -44,8 +44,8 @@
     "apps/api/src/agent-tasks.ts": 1135,
     "apps/api/src/agent-token.test.ts": 16,
     "apps/api/src/agent-token.ts": 638,
-    "apps/api/src/app.test.ts": 386,
-    "apps/api/src/app.ts": 3941,
+    "apps/api/src/app.test.ts": 222,
+    "apps/api/src/app.ts": 3764,
     "apps/api/src/apple-account.test.ts": 119,
     "apps/api/src/apple-account.ts": 383,
     "apps/api/src/apple-auth.test.ts": 182,
@@ -139,7 +139,6 @@
     "apps/api/src/gate-trigger.ts": 1031,
     "apps/api/src/gcs-sign.ts": 207,
     "apps/api/src/genai.ts": 148,
-    "apps/api/src/generator.ts": 38,
     "apps/api/src/github-client.test.ts": 828,
     "apps/api/src/github-client.ts": 4736,
     "apps/api/src/github-rate-limit.ts": 92,
@@ -346,7 +345,6 @@
     "apps/web/src/AdminJobsPanel.test.tsx": 146,
     "apps/web/src/AdminJobsPanel.tsx": 587,
     "apps/web/src/agentActivity.ts": 55,
-    "apps/web/src/api.ts": 46,
     "apps/web/src/App.anonymousPlay.test.ts": 40,
     "apps/web/src/App.catalog.test.ts": 110,
     "apps/web/src/App.createGameFocus.test.ts": 33,
@@ -611,8 +609,6 @@
     "apps/world/src/server.ts": 240,
     "apps/world/src/sim-source.ts": 400,
     "apps/world/src/snapshot-store.ts": 224,
-    "packages/game-generator/src/mock.ts": 118,
-    "packages/game-generator/src/types.ts": 86,
     "packages/zone-core/src/cage.test.ts": 720,
     "packages/zone-core/src/cage.ts": 1441,
     "packages/zone-core/src/contract.test.ts": 251,
