+++ eslint-rules/module-size-baseline.json
@@ -62,8 +62,8 @@
     "apps/api/src/agent-token.ts": 313,
     "apps/api/src/agent-upload-token.test.ts": 87,
     "apps/api/src/agent-upload-token.ts": 164,
-    "apps/api/src/app.test.ts": 561,
-    "apps/api/src/app.ts": 1150,
+    "apps/api/src/app.test.ts": 347,
+    "apps/api/src/app.ts": 1062,
     "apps/api/src/apple-account.test.ts": 137,
     "apps/api/src/apple-account.ts": 81,
     "apps/api/src/apple-auth.test.ts": 203,
@@ -193,7 +193,6 @@
     "apps/api/src/gcs-sign.test.ts": 62,
     "apps/api/src/gcs-sign.ts": 193,
     "apps/api/src/genai.ts": 47,
-    "apps/api/src/generator.ts": 13,
     "apps/api/src/github-client.test.ts": 1944,
     "apps/api/src/github-client.ts": 2084,
     "apps/api/src/github-rate-limit.ts": 19,
@@ -261,7 +260,7 @@
     "apps/api/src/mcp-ui-flag.test.ts": 47,
     "apps/api/src/mcp-ui.test.ts": 611,
     "apps/api/src/mcp-ui.ts": 1874,
-    "apps/api/src/moderation-metrics.test.ts": 175,
+    "apps/api/src/moderation-metrics.test.ts": 174,
     "apps/api/src/moderation-metrics.ts": 86,
     "apps/api/src/moderation-terms.ts": 120,
     "apps/api/src/moderation.test.ts": 228,
@@ -480,7 +479,6 @@
     "apps/web/src/AdminJobsPanel.test.tsx": 344,
     "apps/web/src/AdminJobsPanel.tsx": 462,
     "apps/web/src/agentActivity.ts": 20,
-    "apps/web/src/api.ts": 30,
     "apps/web/src/App.anonymousPlay.test.ts": 123,
     "apps/web/src/App.bottomCta.test.ts": 101,
     "apps/web/src/App.catalog.test.ts": 490,
@@ -489,7 +487,7 @@
     "apps/web/src/App.gamePageTheater.test.ts": 248,
     "apps/web/src/App.overlayExit.test.ts": 161,
     "apps/web/src/App.qa.test.ts": 484,
-    "apps/web/src/App.tsx": 1382,
+    "apps/web/src/App.tsx": 1328,
     "apps/web/src/AppleSignInButton.test.tsx": 120,
     "apps/web/src/AppleSignInButton.tsx": 149,
     "apps/web/src/AppLoadingScreen.tsx": 18,
@@ -563,8 +561,8 @@
     "apps/web/src/CostsPanel.tsx": 171,
     "apps/web/src/createInitialPrompt.test.ts": 18,
     "apps/web/src/createInitialPrompt.ts": 4,
-    "apps/web/src/CreatePage.test.tsx": 117,
-    "apps/web/src/CreatePage.tsx": 140,
+    "apps/web/src/CreatePage.test.tsx": 114,
+    "apps/web/src/CreatePage.tsx": 131,
     "apps/web/src/CreationLimitsPanel.test.tsx": 398,
     "apps/web/src/CreationLimitsPanel.tsx": 411,
     "apps/web/src/creatorGames.test.ts": 75,
@@ -636,8 +634,8 @@
     "apps/web/src/handVerbs.ts": 142,
     "apps/web/src/headerMascot.test.ts": 94,
     "apps/web/src/healthApi.ts": 274,
-    "apps/web/src/HeroPromptSection.test.tsx": 506,
-    "apps/web/src/HeroPromptSection.tsx": 585,
+    "apps/web/src/HeroPromptSection.test.tsx": 475,
+    "apps/web/src/HeroPromptSection.tsx": 578,
     "apps/web/src/howToPlay.test.ts": 257,
     "apps/web/src/howToPlay.ts": 192,
     "apps/web/src/HowToPlayPanel.test.tsx": 221,
@@ -920,11 +918,13 @@
     "packages/contract/src/delivery-mode.ts": 3,
     "packages/contract/src/dismiss-reason.test.ts": 8,
     "packages/contract/src/dismiss-reason.ts": 15,
+    "packages/contract/src/game-project.test.ts": 15,
+    "packages/contract/src/game-project.ts": 8,
     "packages/contract/src/gate-progress.test.ts": 8,
     "packages/contract/src/gate-progress.ts": 3,
     "packages/contract/src/gate-status.test.ts": 49,
     "packages/contract/src/gate-status.ts": 25,
-    "packages/contract/src/index.ts": 110,
+    "packages/contract/src/index.ts": 111,
     "packages/contract/src/job-state.test.ts": 35,
     "packages/contract/src/job-state.ts": 47,
     "packages/contract/src/managed-agent-vendor.test.ts": 8,
@@ -951,12 +951,6 @@
     "packages/contract/src/zone-link-step.ts": 4,
     "packages/contract/src/zone-protocol.test.ts": 8,
     "packages/contract/src/zone-protocol.ts": 2,
-    "packages/game-generator/src/engine/engine.test.ts": 35,
-    "packages/game-generator/src/index.ts": 2,
-    "packages/game-generator/src/mock.test.ts": 67,
-    "packages/game-generator/src/mock.ts": 102,
-    "packages/game-generator/src/types.ts": 23,
-    "packages/game-generator/vitest.config.ts": 8,
     "packages/zone-core/src/cage.test.ts": 317,
     "packages/zone-core/src/cage.ts": 464,
     "packages/zone-core/src/contract.test.ts": 123,
