+++ packages/contract/src/game-project.ts
@@ -1,0 +1,8 @@
+// One game's source, before assembly into a single document.
+export interface GameProject {
+  title: string;
+  description: string;
+  html: string;
+  js: string;
+  css: string;
+}
