--- packages/game-generator/package.json
@@ -1,17 +1,0 @@
-{
-  "name": "@gamedevpl/game-generator",
-  "version": "0.0.0",
-  "private": true,
-  "type": "module",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "scripts": {
-    "build": "tsc -p tsconfig.json",
-    "type-check": "tsc -p tsconfig.json --noEmit",
-    "test": "vitest run"
-  },
-  "devDependencies": {
-    "typescript": "^5.5.4",
-    "vitest": "^3.2.7"
-  }
-}
