--- packages/game-generator/src/index.ts
@@ -1,2 +1,0 @@
-export * from './types.js';
-export * from './mock.js';
