--- packages/game-generator/templates/collect/index.html
@@ -1,6 +1,0 @@
-<div class="wrap">
-  <h1 id="game-title">__TITLE__</h1>
-  <p id="game-desc">__DESCRIPTION__</p>
-  <canvas id="game" width="480" height="640"></canvas>
-  <p class="hint">Move the basket with the LEFT / RIGHT arrow keys and catch the falling coins.</p>
-</div>
