--- packages/game-generator/templates/dodge/style.css
@@ -1,30 +1,0 @@
-.wrap {
-  font-family: system-ui, sans-serif;
-  text-align: center;
-  color: #e8e8f0;
-}
-
-.wrap h1 {
-  margin: 0.2em 0 0.1em;
-  font-size: 1.4rem;
-}
-
-.wrap #game-desc {
-  margin: 0 0 0.6em;
-  opacity: 0.75;
-  font-size: 0.9rem;
-}
-
-#game {
-  background: #10121c;
-  border: 2px solid #2c2f45;
-  border-radius: 8px;
-  max-width: 100%;
-  touch-action: none;
-}
-
-.hint {
-  margin-top: 0.6em;
-  font-size: 0.8rem;
-  opacity: 0.6;
-}
