--- packages/game-generator/templates/space/index.html
@@ -1,6 +1,0 @@
-<div class="wrap">
-  <h1 id="game-title">__TITLE__</h1>
-  <p id="game-desc">__DESCRIPTION__</p>
-  <canvas id="game" width="480" height="640"></canvas>
-  <p class="hint">Fly with the arrow keys or WASD — you wrap around the edges. Dodge the asteroids to survive.</p>
-</div>
